+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 - updated Affiliation Schema, Resolver, Models to use new affiliations tables in the database
 - updated all of the cache key structures to wrap them in `{}` due to the way Redis handles keys in cluster mode
 - updated emailService to use nodemailer and to support emailConfirmation templateCollaboration and planCollaboration email messages
+- added bestPractice flag to the Section
 
 ### Fixed
 - Converted DateTimeISO to String in schemas so that dates could be inserted into mariaDB database, and updated MySqlModel and associated unit test